requirements.txt
streamlit
pandas
numpy
openpyxl
plotly
langchain
langchain-community
langchain-openai
faiss-cpu
openai
python-dotenv